+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
 **Improvements**
  - public_model task also defines methods mapped to MAS module steps when publishing to MAS.
  - SSL verification can be disable with `SSLREQCERT` environment variable.
+ - CAs to use for validating SSL certificates can also be specified through the `SSLCALISTLOC` environment variable.
  - Added `execute_performance_task`
 
 **Changes**